Move `Antialiasing` to `iced_graphics`
