+++ glow/src/settings.rs
@@ -1,6 +1,7 @@
 //! Configure a [`Renderer`].
 //!
 //! [`Renderer`]: struct.Renderer.html
+pub use iced_graphics::Antialiasing;
 
 /// The settings of a [`Renderer`].
 ///
@@ -21,30 +22,6 @@ impl Default for Settings {
         Settings {
             default_font: None,
             antialiasing: None,
-        }
-    }
-}
-
-/// An antialiasing strategy.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Antialiasing {
-    /// Multisample AA with 2 samples
-    MSAAx2,
-    /// Multisample AA with 4 samples
-    MSAAx4,
-    /// Multisample AA with 8 samples
-    MSAAx8,
-    /// Multisample AA with 16 samples
-    MSAAx16,
-}
-
-impl Antialiasing {
-    pub(crate) fn sample_count(self) -> u32 {
-        match self {
-            Antialiasing::MSAAx2 => 2,
-            Antialiasing::MSAAx4 => 4,
-            Antialiasing::MSAAx8 => 8,
-            Antialiasing::MSAAx16 => 16,
         }
     }
 }
